Trait http_util::header::Encoded

pub trait Encoded: for<'a> Iterable<'a> {
    fn is_empty(&self) -> bool { ... }
}

An encoded header value.

A single header value can be represented in multiple fields. In this case, the value is equivalent to a comma separated list of all the fields.

Provided Methods

fn is_empty(&self) -> bool

Returns true if the encoded header value is empty.